+++ CHANGELOG.md
@@ -278,6 +278,7 @@
 * [X] breeze-core 将`mapOfValueNotNull`重命名为`mapOfValuesNotNull`。
 * [X] breeze-http&breeze-logger 重构代码。
 * [X] breeze-core 完善对html和xml的转义逻辑。（来自guava）
+* [X] breeze-core 优化`String.replaceAll`和`String.escapeBy`等方法。
 * [ ] breeze-core 让集合的`deepQuery`方法当特定的占位符不匹配类型时，回调为普通字符串。
 * [ ] breeze-core 从类型推断出对应的默认值。（来自guava）
 * [ ] breeze-unstable 新的集合类型：`Tree`。